refactor: объединить character_builder.py в character_build.py
- Объединить содержимое character_builder.py в character_build.py
- Удалить character_builder.py
- Обновить все импорты из character_builder на character_build
- Использовать lazy imports для избежания circular import
- Тесты проходят
