+++ tests/test_saving_throws.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 import pytest
 
-from core.character_builder import resolve_creation_grants
+from core.character_build import resolve_creation_grants
 from core.checks import saving_throw, saving_throw_modifier
 from core.models import Character
 from core.proficiencies.proficiency_checks import get_class_saving_throws
